refactor: Update application configuration and remove unused API platform config

+++ packages/Webkul/CustomerVerification/src/Database/Migrations/2026_04_17_add_admin_verification_tables.php
@@ -17,7 +17,7 @@ return new class extends Migration
 
         if (! Schema::hasTable('verification_audit_logs')) Schema::create('verification_audit_logs', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('admin_id')->nullable();
+            $table->unsignedInteger('admin_id')->nullable();
             $table->unsignedBigInteger('customer_id');
             $table->string('action', 50); // approved, rejected, viewed
             $table->text('reason')->nullable();
